requirements.txt
pdfplumber>=0.10.3
camelot-py>=0.11.0
pytesseract>=0.3.10
pillow>=10.2.0
python-dotenv>=1.0.0
msgpack>=1.0.7
